Refactor AccountConfig component by extracting form validation logic and related types into a new accountConfigForm module, improving code organization and maintainability
